Limit petz spawning to avoid animal overload
The number of petz that can spawn in a certain area has been reduced
from the default 60, to avoid massive animal herding when staying longer
at a certain place.  The type of animals that can spawn has also been
reduced to avoid some animals that can attack players – because this is
annoying.